Correct the size of _Unwind_Exception.
To fix the alias1 exception test for clang on MIPS architecture, and
increase the compatibility with gcc, we should use uint32_t for
private_1 and private_2.

With this change, the sizeof(_Unwind_Exception) == 24 on MIPS,
and sizeof(_Unwind_Exception) == 32 on X86.

+++ sources/cxx-stl/gabi++/include/unwind-itanium.h
@@ -68,8 +68,8 @@ typedef _Unwind_Reason_Code (*_Unwind_Stop_Fn) (int version,
 struct _Unwind_Exception {
   uint64_t exception_class;
   _Unwind_Exception_Cleanup_Fn exception_cleanup;
-  uint64_t private_1;
-  uint64_t private_2;
+  uint32_t private_1;
+  uint32_t private_2;
 } __attribute__((__aligned__)); // must be double-word aligned
 
 _Unwind_Reason_Code _Unwind_RaiseException(struct _Unwind_Exception*);
